Implemented code review suggestions, cleaned the code and made the code more streamlined.
