+++ IS7024_01_23/Pages/Index.cshtml
@@ -6,10 +6,9 @@
 @model IndexModel
 @{
     ViewData["Title"] = "Home page";
-    var parks = ViewData["Parks"];
+    // var parks = ViewData["Parks"];
 }
 
-<link rel="stylesheet" href="~/css/styles.css" />
 
 <div class="container text-center centerdiv">
     <h2 class="headformat">EXPLORE AMERICA'S HIKING TRAILS</h2>
@@ -27,20 +26,6 @@
     </form>
 </div>
 
-@* <div class="container text-center centerdiv">
-    <h2 class="headformat">EXPLORE AMERICA'S HIKING TRAILS</h2>
-    <form method="post" class="center-form">
-        <label for="stateDropdown" class="sizeup">Select a State:</label>
-        <select id="stateDropdown" name="stateCode" class="sizeup">
-            @foreach (var state in GetStatesFromJsonFile("states.json"))
-            {
-                <option value=@state.Value>@state.Key</option>
-            }
-            }
-        </select>
-        <button type="submit" class="sizeup"><a class="nav-link text-dark" asp-area="" asp-page="/ViewParks">Submit</a></button>
-    </form>
-</div> *@
 <script>
     document.getElementById('parkForm').addEventListener('submit', function (e) {
         e.preventDefault(); // Prevent the default form submission
