+++ IS7024_01_23/Pages/ParkDetail.cshtml
@@ -13,7 +13,7 @@
     <!-- Add other fields here as needed -->
 </div>
 
-<script>
+@* <script>
     document.addEventListener("DOMContentLoaded", function () {
         const urlParams = new URLSearchParams(window.location.search);
         const parkName = urlParams.get("parkName");
@@ -31,4 +31,4 @@
         document.getElementById("park-image").src = image;
         // You can populate other fields in a similar way
     });
-</script>
+</script> *@
