+++ IS7024_01_23/ParksData.cs
@@ -5,16 +5,15 @@
 //    using parksNamespace;
 //
 //    var parksData = ParksData.FromJson(jsonString);
+using System;
+using System.Collections.Generic;
+
+using System.Globalization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 
 namespace parksNamespace
 {
-    using System;
-    using System.Collections.Generic;
-
-    using System.Globalization;
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Converters;
-
     public partial class ParksData
     {
         [JsonProperty("parks")]
@@ -25,6 +24,9 @@ namespace parksNamespace
     {
         [JsonProperty("parkName")]
         public string ParkName { get; set; }
+
+        [JsonProperty("parkID")]
+        public string ParkID{ get; set; }
 
         [JsonProperty("description")]
         public string Description { get; set; }
