+++ IS7024_01_23/WeatherData.cs
@@ -1,12 +1,12 @@
-﻿namespace WeatherSpace
+﻿using System;
+using System.Collections.Generic;
+
+using System.Globalization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace WeatherSpace
 {
-    using System;
-    using System.Collections.Generic;
-
-    using System.Globalization;
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Converters;
-
     public partial class WeatherData
     {
         [JsonProperty("city_name")]
